+++ frontend/src/components/jobs/filters/dropdown-sort.tsx
@@ -1,5 +1,5 @@
 import { Select } from "@mantine/core";
-import { useFilterContext } from "@/context/jobs/filter-context";
+import { useFilterContext } from "@/context/filter/filter-context";
 
 export default function DropdownSort() {
   const { filters } = useFilterContext();
